requirements txt.txt
livekit-agents>=0.12.0
livekit-plugins-google>=0.10.0
livekit-plugins-silero>=0.7.0
livekit-plugins-simli>=0.2.0
python-dotenv>=1.0.0
